Move `NamespaceSocketStore` and `NamespaceStore` to their own files
TLDR; The below description is complex, so it is better to have a simple
example:

Suppose you created a server named `io`. Normally, you would access
`NamespaceSocketStore` by doing `io.Of("/").SocketStore()`, but after this commit, you cannot. Only the `Namespace` and adapter should access it.

Long explanation:

Improved the API of adapter by adding `socketStore` parameter to `newInMemoryAdapter`. Now we access socket store directly from the adapter — instead of accessing it from `Namespace` —, and this leads to a cleaner API. If this wouldn't be done, the user (not the adapter developer, socket.io-go user) would access the `NamespaceSocketStore` directly from the `Namespace`, which is a bad design.

Removed leftover `serverConnStore`